Move Article classes to Domain Article
