+++ app/Domain/Article/ArticleDTO.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Dto;
+namespace App\Domain\Article;
 
 class ArticleDTO
 {
