+++ app/Domain/Article/ArticlePolicy.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Policies;
+namespace App\Domain\Article;
 
 use App\Models\User;
