+++ app/Http/Controllers/Api/ArticleController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers\Api;
 
-use App\Contracts\ArticleServiceInterface;
+use App\Domain\Article\Contracts\ArticleServiceInterface;
 use Illuminate\Http\JsonResponse;
 
 class ArticleController
