+++ app/Providers/AppServiceProvider.php
@@ -2,16 +2,16 @@
 
 namespace App\Providers;
 
-use App\Contracts\ArticleRepositoryInterface;
-use App\Contracts\ArticleServiceInterface;
 use App\Contracts\CategoryRepositoryInterface;
 use App\Contracts\CategoryServiceInterface;
 use App\Contracts\CommentRepositoryInterface;
 use App\Contracts\CommentServiceInterface;
-use App\Repository\ArticleRepository;
+use App\Domain\Article\Contracts\ArticleRepositoryInterface;
+use App\Domain\Article\Contracts\ArticleServiceInterface;
+use App\Domain\Article\Repository\ArticleRepository;
+use App\Domain\Article\Services\ArticleService;
 use App\Repository\CategoryRepository;
 use App\Repository\CommentRepository;
-use App\Services\ArticleService;
 use App\Services\CategoryService;
 use App\Services\CommentService;
 use Illuminate\Support\ServiceProvider;
